Employee and Signup

+++ db/migrate/20220810181414_create_employees.rb
@@ -3,7 +3,7 @@ class CreateEmployees < ActiveRecord::Migration[6.1]
     create_table :employees do |t|
     t.string :name
     t.string :email
-    t.integer :phonenumber
+    t.bigint :phonenumber
     t.string :emp_id
     t.string :devices
     t.string :role
